Add use it button

+++ src/Items/HpPotion.cpp
@@ -6,6 +6,6 @@ HpPotion::HpPotion(Context* context) : Pickable(context)
 {
 }
 
-const int &HpPotion::get_hp_points_to_restore() const {
+const float &HpPotion::get_hp_points_to_restore() const {
     return m_hp_points;
 }
